+++ apps/evalite-ui/app/routeTree.gen.ts
@@ -27,13 +27,15 @@ const EvalNameRoute = EvalNameImport.update({
   id: '/eval/$name',
   path: '/eval/$name',
   getParentRoute: () => rootRoute,
-} as any)
+} as any).lazy(() => import('./routes/eval.$name.lazy').then((d) => d.Route))
 
 const EvalNameResultResultIndexRoute = EvalNameResultResultIndexImport.update({
   id: '/result/$resultIndex',
   path: '/result/$resultIndex',
   getParentRoute: () => EvalNameRoute,
-} as any)
+} as any).lazy(() =>
+  import('./routes/eval.$name.result.$resultIndex.lazy').then((d) => d.Route),
+)
 
 // Populate the FileRoutesByPath interface
